pyproject.toml
[project]
name = "findmyrent"
version = "0.1.0"
description = "Backend API for the FindMyRent application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.0",
    "beanie>=2.0.0",
    "clicksend-client>=5.0.78",
    "cloudinary>=1.44.1",
    "fastapi-limiter>=0.1.6",
    "fastapi-middlewares>=0.1.0",
    "fastapi[standard]>=0.116.1",
    "filetype>=1.2.0",
    "httpx>=0.28.1",
    "logfire[fastapi,system-metrics]>=4.14.2",
    "motor>=3.7.1",
    "opentelemetry-instrumentation-httpx>=0.59b0",
    "opentelemetry-instrumentation-pymongo>=0.59b0",
    "opentelemetry-instrumentation-redis>=0.59b0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "passlib>=1.7.4",
    "psycopg>=3.2.10",
    "psycopg2>=2.9.10",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
]
